Running Version

+++ src/main/java/com/example/projectdb/repo/HawkerListingRepository.java
@@ -3,9 +3,7 @@ package com.example.projectdb.repo;
 import com.example.projectdb.model.HawkerListing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HawkerListingRepository extends JpaRepository<HawkerListing, Integer>{
-
-
+public interface HawkerListingRepository extends JpaRepository<HawkerListing, Long>{
 
 
 }
